fix: optimize
